python_server/requirements.txt
fastapi
uvicorn[standard]
python-multipart
PyPDF2
python-docx
motor
pymongo
groq
pydantic-settings
httpx
python-dotenv
pdfplumber
filetype
aiofiles
loguru
nltk
typing-extensions
magic-python
pydantic